requirements.txt
absl-py==0.7.1
astor==0.8.0
decorator==4.4.0
dgl==0.3
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
networkx==2.3
numpy==1.16.2
Pillow==5.4.1
protobuf==3.8.0
PyYAML==3.13
scipy==1.3.0
six==1.12.0
tensorboard==1.13.0
tensorboardX==1.6
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
torch==1.0.1
-e git+https://github.com/rtqichen/torchdiffeq.git@8c6078974786bb4ca879fc35ac74edc27aa82a19#egg=torchdiffeq
torchvision==0.2.2
tqdm==4.19.9
Werkzeug==0.15.4
